Lecture 88, storing session in database
It is a matter of configuring, for the most part.
It is also necessary to create a 'sessions' table in the db.
In order to do so, we need to create a manual migration (migrations:generate vs migrations:diff)
and populate it with the proper sql create table query.
Now, as it happens, the next migrations:diff is going to find out that there are discrepancies btw tables and migrations and entities, and it is going to try to remove the sessions table.
In order to prevent so, we add a new config line in doctrine.yaml